added BookDetails class and some relationships
